Lead-Gen-website-Chatbot-main/lead-gen-chatbot-final/requirements.txt
annotated-types
anyio
beautifulsoup4
certifi
charset-normalizer
click
dnspython
email_validator
exceptiongroup
fastapi
fastapi-cli
h11
httpcore
httptools
httpx
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
pydantic
pydantic_core
Pygments
python-dotenv
python-multipart
PyYAML
requests
rich
rich-toolkit
shellingham
sniffio
soupsieve
starlette
typer
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
aiohappyeyeballs
aiohttp
aiosignal
airtable
annotated-types
anyio
asgiref
async-timeout
attrs
backoff
bcrypt
beautifulsoup4
build
cachetools
certifi
charset-normalizer
chroma-hnswlib
chromadb
click
coloredlogs
dataclasses-json
Deprecated
distro
dnspython
durationpy
email_validator
exceptiongroup
fastapi
fastapi-cli
filelock
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
greenlet
groq
grpcio
h11
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
Jinja2
joblib
jsonpatch
jsonpointer
kubernetes
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pillow
posthog
propcache
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
Pygments
PyPika
pyproject_hooks
python-dateutil
python-dotenv
python-multipart
PyYAML
regex
requests
requests-oauthlib
requests-toolbelt
rich
rich-toolkit
rsa
safetensors
scikit-learn
scipy
sentence-transformers
shellingham
six
sniffio
soupsieve
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tokenizers
tomli
torch
tqdm
transformers
triton
typer
typing-inspect
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websocket-client
websockets
wrapt
yarl
zipp
